+++ srcs/config/read_config.cpp
@@ -1,6 +1,6 @@
 #include "ServerConfig.hpp"
+#include "utils/file_io_utils.hpp"
 #include "utils/tokenize.hpp"
-#include "utils/util.hpp"
 #include <cstdlib>
 #include <iostream>
 #include <vector>
