+++ srcs/http/method.cpp
@@ -3,7 +3,8 @@
 #include "const_html_filename.hpp"
 #include "const_response_key_map.hpp"
 #include "const_status_phrase.hpp"
-#include "utils/util.hpp"
+#include "utils/file_io_utils.hpp"
+#include "utils/utils.hpp"
 #include <string>
 #include <unistd.h>
